Extract `state_dir`
