+++ src/tribler/core/components/libtorrent/conftest.py
@@ -1,6 +1,8 @@
 import pytest
 
 from tribler.core.components.libtorrent.download_manager.download_config import DownloadConfig
+from tribler.core.components.libtorrent.torrentdef import TorrentDef
+from tribler.core.tests.tools.common import TESTS_DATA_DIR
 
 
 # pylint: disable=redefined-outer-name
@@ -9,3 +11,14 @@ from tribler.core.components.libtorrent.download_manager.download_config import 
 @pytest.fixture
 def download_config():
     return DownloadConfig()
+
+
+@pytest.fixture
+def test_tdef(state_dir):
+    tdef = TorrentDef()
+    sourcefn = TESTS_DATA_DIR / 'video.avi'
+    tdef.add_content(sourcefn)
+    tdef.set_tracker("http://localhost/announce")
+    torrentfn = state_dir / "gen.torrent"
+    tdef.save(torrentfn)
+    return tdef
